refactor: split up the upload endpoints
