+++ server/routes/upload/getActiveJobs.ts
@@ -1,0 +1,16 @@
+import { Request, Response } from "express";
+
+import ConversionJob from "../../lib/jobs/ConversionJob";
+import DB from "../../lib/storage/db";
+
+export default async function getActiveJobs(req: Request, res: Response) {
+  console.debug("getting jobs");
+  try {
+    ConversionJob.ResumeStaleJobs(DB, res.locals.owner);
+    let c = new ConversionJob(DB);
+    const jobs = await c.AllStartedJobs(res.locals.owner);
+    res.send(jobs);
+  } catch (error) {
+    console.error(error);
+  }
+}
